metricEndpoint is typescripted but untested
