+++ metricEndpoint/src/proto/merged/metricMerged.proto
@@ -1,16 +1,5 @@
 syntax = "proto3";
 
-message Any {
-  string type_url = 1;
-  bytes value = 2;
-}
-
-message Status {
-  int32 code = 1;
-  string message = 2;
-  repeated Any details = 3;
-}
-
 message AnyValue {
   oneof value {
     string string_value = 1;
@@ -204,89 +193,5 @@ message ExportMetricsServiceResponse {
 
 message ExportMetricsPartialSuccess {
   int64 rejected_data_points = 1;
-  string error_message = 2;
-}
-
-message TracesData {
-  repeated ResourceSpans resource_spans = 1;
-}
-
-message ResourceSpans {
-  reserved 1000;
-  Resource resource = 1;
-  repeated ScopeSpans scope_spans = 2;
-  string schema_url = 3;
-}
-
-message ScopeSpans {
-  InstrumentationScope scope = 1;
-  repeated Span spans = 2;
-  string schema_url = 3;
-}
-
-message Span {
-  bytes trace_id = 1;
-  bytes span_id = 2;
-  string trace_state = 3;
-  bytes parent_span_id = 4;
-  string name = 5;
-  enum SpanKind {
-    SPAN_KIND_UNSPECIFIED = 0;
-    SPAN_KIND_INTERNAL = 1;
-    SPAN_KIND_SERVER = 2;
-    SPAN_KIND_CLIENT = 3;
-    SPAN_KIND_PRODUCER = 4;
-    SPAN_KIND_CONSUMER = 5;
-  }
-  SpanKind kind = 6;
-  fixed64 start_time_unix_nano = 7;
-  fixed64 end_time_unix_nano = 8;
-  repeated KeyValue attributes = 9;
-  uint32 dropped_attributes_count = 10;
-  message Event {
-    fixed64 time_unix_nano = 1;
-    string name = 2;
-    repeated KeyValue attributes = 3;
-    uint32 dropped_attributes_count = 4;
-  }
-  repeated Event events = 11;
-  uint32 dropped_events_count = 12;
-  message Link {
-    bytes trace_id = 1;
-    bytes span_id = 2;
-    string trace_state = 3;
-    repeated KeyValue attributes = 4;
-    uint32 dropped_attributes_count = 5;
-  }
-  repeated Link links = 13;
-  uint32 dropped_links_count = 14;
-  TraceStatus status = 15;
-}
-
-message TraceStatus {
-  reserved 1;
-  string message = 2;
-  enum StatusCode {
-    STATUS_CODE_UNSET               = 0;
-    STATUS_CODE_OK                  = 1;
-    STATUS_CODE_ERROR               = 2;
-  };
-  StatusCode code = 3;
-}
-
-service TraceService {
-  rpc Export(ExportTraceServiceRequest) returns (ExportTraceServiceResponse) {}
-}
-
-message ExportTraceServiceRequest {
-  repeated ResourceSpans resource_spans = 1;
-}
-
-message ExportTraceServiceResponse {
-  ExportTracePartialSuccess partial_success = 1;
-}
-
-message ExportTracePartialSuccess {
-  int64 rejected_spans = 1;
   string error_message = 2;
 }
