+++ metricEndpoint/src/proto/merged/statusMerged.proto
@@ -1,0 +1,15 @@
+syntax = "proto3";
+
+message Any {
+  string type_url = 1;
+  bytes value = 2;
+}
+
+message Status {
+  int32 code = 1;
+  string message = 2;
+  repeated Any details = 3;
+}
+
+
+
