+++ metricEndpoint/src/proto/original/any.proto
