+++ metricEndpoint/src/proto/original/common.proto
