+++ metricEndpoint/src/proto/original/metrics_service.proto
