+++ metricEndpoint/src/proto/original/resource.proto
