+++ metricEndpoint/src/proto/original/status.proto
